+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Major update on dependencies
 - Update how to start dev server in https
 - More validation on public function params (type errors thrown on unexpected params)
+- "errored" execution status becomes "failed"
 
 # 30.4.0